% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootpval.R
\name{bootpval}
\alias{bootpval}
\title{Bootstrapping function to find the p-value}
\usage{
bootpval(x, conf.level = 0.95, iter = 3000, mu0 = 0, test = "two")
}
\arguments{
\item{x}{sample}

\item{conf.level}{confidence interval level}

\item{iter}{number of iterations to perform bootstrapping for}

\item{mu0}{the null hypothesis for the mean}

\item{test}{test tailedness selector (two, upper, lower)}
}
\value{
the p-value, the tcalc value, and the confidence interval
}
\description{
Function for finding p-value of one sample t-test
}
\examples{
bootpval(rnorm(30, 0, 1))

}
