% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire.R
\docType{data}
\name{fire}
\alias{fire}
\title{Fire Damage data set.}
\format{A data frame with 15 rows and 2 variables:
\describe{
  \item{DISTANCE}{distance from the nearest fire station, in miles}
  \item{DAMAGE}{damage caused by the fire, in thousands of dollars}
}}
\source{
\url{https://www.crcpress.com/Statistics-for-Engineering-and-the-Sciences/Mendenhall-Sincich/p/book/9781498728850}
}
\usage{
fire
}
\description{
A data set collected by a fire insurance company
}
\keyword{datasets}
