% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2.R
\name{myboot2}
\alias{myboot2}
\title{Bootstrapping function}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{number of samples to take}

\item{x}{sample data}

\item{fun}{function to approximate}

\item{alpha}{error for confidence interval}

\item{cx}{text scale}
}
\value{
a plot of the sample data, along with statistic of interest, point estimate, and the confidence interval
}
\description{
This function uses bootstrapping to approximate properties of a population
}
