% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myci.R
\name{myci}
\alias{myci}
\title{Confidence interval for the mean}
\usage{
myci(data, alpha = 0.05)
}
\arguments{
\item{data}{sample to find interval for}

\item{alpha}{confidence interval variable}
}
\value{
confidence interval
}
\description{
Function for finding the confidence interval of the mean
}
