% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{Normal curve plot function}
\usage{
myncurve(mu = 0, sigma = 1, a = 0)
}
\arguments{
\item{mu}{The mean of the distribution}

\item{sigma}{The standard deviation of the distribution}

\item{a}{The value a for which P(X < a) is calculated}
}
\value{
A plot of the probability density function and the calculation P(X < a)
}
\description{
This function plots a normal probability distribution
and highlights the density for P(X < a)
}
\examples{
myncurve(mu = 10, sigma = 4, a = 6)

}
