% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myNRML.R
\name{myNRML}
\alias{myNRML}
\title{Newton Raphson Method}
\usage{
myNRML(x0, delta = 0.001, llik, xrange, parameter = "param")
}
\arguments{
\item{x0}{initial estimate}

\item{delta}{timestep for the derivative}

\item{llik}{log likelihood function}

\item{xrange}{x range of the graph produced by this function}

\item{parameter}{main title of the graph}
}
\value{
Plot of the
}
\description{
Estimates the likelihood of a parameter
}
