% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obinom.R
\name{obinom}
\alias{obinom}
\title{Occurences in a binomial}
\usage{
obinom(size = 10, p = 0.5)
}
\arguments{
\item{size}{The size of the distribution}

\item{p}{The probability of a trial being a success}
}
\value{
A dataframe with the probabilities of the number of successes for each possible number of successes
}
\description{
This function returns the probability for numbers of successes in a binomial distribution
}
\examples{
obinom()

}
