% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotModel.R
\name{plotLM}
\alias{plotLM}
\title{Straight line model plotting function}
\usage{
plotLM(data, x_val, y_val, col, title)
}
\arguments{
\item{data}{Data to plot}

\item{x_val}{Variable for the x axis}

\item{y_val}{Variable for the y axis}

\item{col}{Color of the plot}

\item{title}{Title of the plot}
}
\value{
A plot of the linear model
}
\description{
Plots a straight line linear model with scatter plot
}
\examples{
plotLM(spruce.df, spruce.df$BHDiameter, spruce.df$Height, spruce.df$BHDiameter, "Height vs Diameter")


}
