% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.R
\name{plotOutliers}
\alias{plotOutliers}
\title{Outlier plot function}
\usage{
plotOutliers(x, title = "", col = c("Red", "Blue", "Black"))
}
\arguments{
\item{x}{The data to plot}

\item{col}{Color vector containing outlier color, potential outlier color, and regular color (in that order)}
}
\value{
Plots a dot plot of the data points, highlighting outliers
}
\description{
This function plots data points, outliers, and potential outliers in a dot plot
}
