% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quadraticTrendline.R
\name{quadraticTrendline}
\alias{quadraticTrendline}
\title{Quadratic trendline function}
\usage{
quadraticTrendline(model)
}
\arguments{
\item{model}{Linear model for finding the coefficients}
}
\value{
An equation for the quadratic trendline
}
\description{
Finds the equation for a quadratic trendline
}
