% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/violin.R
\name{violin}
\alias{violin}
\title{Violin plot function}
\usage{
violin(data, x_val, y_val, fill_col, title)
}
\arguments{
\item{data}{The data frame to plot with}

\item{x_val}{The x axis values for the violin plots}

\item{y_val}{The y axis values for the violin plots}

\item{fill_col}{Variable to base the fill color on}

\item{title}{Name for the plot}
}
\value{
A violin plot of the data
}
\description{
This function makes violin plots using ggplot2
}
\examples{
violin(ddt$RIVER, ddt$LENGTH, ddt$SPECIES, "Violin Plot")

}
